==> Sentiment/3. origin_bbox_plot.ipynb <==
import os
import numpy as np
import re
from PIL import Image, ImageOps, UnidentifiedImageError
import pandas as pd
import cv2

fil2=pd.read_csv('./merged_df.csv')
fil2

file_name=[]
original_width=[]
original_height=[]
resize_ratio=[]
padding_x=[]
padding_y=[]



main_dir='.\\원천데이터'

emo_list=os.listdir(main_dir)

emo_list
for emo in emo_list: # 감성 리스트
    file_folder_path=os.path.join(main_dir,emo) # main_dir + 감성 
    print(file_folder_path)


    match=re.split('_', emo) # 폴더명에서 감성 (기쁨, 슬픔 ) 만 가져오기
    match=match[1]

    file_list=os.listdir(file_folder_path) # 파일이 들어있는 폴더 경로
    print(file_list)

    for num, file in enumerate(file_list):
        origin_file_adress=os.path.join(file_folder_path,file) #파일 경로

        file_name.append(file) # 파일명 저장

        if os.path.exists(origin_file_adress): # 파일이 확인되면

            try:
                origin_img = Image.open(origin_file_adress) # 파일을 가져와서
                img = ImageOps.exif_transpose(origin_img)

                bounding_box=fil2.iloc[num]
                x0 = bounding_box['bbox_x_center_list'] - bounding_box['bbox_widthlist'] / 2
                x1 = bounding_box['bbox_x_center_list'] + bounding_box['bbox_widthlist'] / 2
                y0 = bounding_box['bbox_y_center_list'] - bounding_box['bbox_heightlist'] / 2
                y1 = bounding_box['bbox_y_center_list'] + bounding_box['bbox_heightlist'] / 2
                                # print('x0=',x0, x1, 'y0=',y0, y1)
                start_point = (int(x0), int(y0))
                end_point = (int(x1), int(y1))
                image = cv2.rectangle(np.array(img), start_point, end_point, color=(0, 255, 0), thickness=2)

            
                cv2.imwrite(f"./aabox/bounding_boxes{num}.jpg", image)
                print(origin_file_adress)


            except (OSError, UnidentifiedImageError) as e: #이미지 손상등 에러가 있으면 넘어간다
                if 'image file is truncated' in str(e):
                    print(f"Error loading image: {origin_file_adress}. Skipping to the next image.")
                    continue
                else:
                    raise e
