==> Sentiment/3. resize_bbox_plot.ipynb <==
import os
import numpy as np
import re
import json
import pandas as pd
from PIL import Image, ImageOps, UnidentifiedImageError


# 데이터프레임 두개 merge
fil1=pd.read_csv('./orizin_bbox_dot.csv')
fil2=pd.read_csv('./orizin_resize.csv')

merge_df = pd.merge(fil1, fil2, on='file_name', how='left')
merge_df.to_csv('./merged_df.csv',index=False)
merge_df

for index, row in merge_df.iterrows():
    resize_x_center = (row['bbox_x_center_list']) * row['resize_width_ratio']+row['padding_x']
    resize_y_center = row['bbox_y_center_list'] * row['resize_width_ratio']  +row['padding_y']
    resize_width = row['bbox_widthlist'] * row['resize_width_ratio']
    resize_height = row['bbox_heightlist'] * row['resize_width_ratio']
    
    merge_df.loc[index, 'resize_x_center'] = resize_x_center
    merge_df.loc[index, 'resize_y_center'] = resize_y_center
    merge_df.loc[index, 'resize_width'] = resize_width
    merge_df.loc[index, 'resize_height'] = resize_height


resized_df=merge_df[['file_name','resize_x_center','resize_y_center','resize_width','resize_height']]
resized_df.to_csv('./resized_df.csv',index=False)
resized_df


import os
from PIL import Image, ImageOps
import cv2

main_dir = '.\\수정본'
emo_list = os.listdir(main_dir)

for emo in emo_list: # 감성 리스트
    file_folder_path = os.path.join(main_dir, emo) # main_dir + 감성 
    print(emo)

    file_list = os.listdir(file_folder_path)
    for num,file in enumerate(file_list):
        file_path = os.path.join(file_folder_path, file)
        if os.path.exists(file_path): # 파일이 확인되면
            try:
                origin_img = Image.open(file_path) # 파일을 가져와서
                img = ImageOps.exif_transpose(origin_img)

                bounding_box = resized_df[resized_df['file_name'] == file]

                print(file)
                x0 = bounding_box['resize_x_center'].values[0] - bounding_box['resize_width'].values[0] / 2
                x1 = bounding_box['resize_x_center'].values[0] + bounding_box['resize_width'].values[0] / 2
                y0 = bounding_box['resize_y_center'].values[0] - bounding_box['resize_height'].values[0] / 2
                y1 = bounding_box['resize_y_center'].values[0] + bounding_box['resize_height'].values[0] / 2
                
          
                print('x0=',x0, x1, 'y0=',y0, y1)
                start_point = (int(x0), int(y0))
                end_point = (int(x1), int(y1))
                image = cv2.rectangle(np.array(img), start_point, end_point, color=(0, 255, 0), thickness=1)
                image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
                cv2.imwrite(f"./box/{num}.jpg", image)

            except (OSError, UnidentifiedImageError) as e: # 이미지 손상 등 에러가 있으면 넘어갑니다.
                if 'image file is truncated' in str(e):
                    print(f"Error loading image: {file_path}. Skipping to the next image.")
                    continue
                else:
                    raise e


matching_rows = resized_df[resized_df['file_name'] == file]
aa=matching_rows['resize_x_center'].values[0]
bounding_box=resized_df.iloc[num]
bounding_box

bounding_box['resize_width'].values[0]

import os
from PIL import Image, ImageOps
import cv2
import time

main_dir = '.\\수정본'
emo_list = os.listdir(main_dir)

for emo in emo_list: # 감성 리스트
    file_folder_path = os.path.join(main_dir, emo) # main_dir + 감성 
    print(emo)

    file_list = os.listdir(file_folder_path)
    for num,file in enumerate(file_list):
        file_path = os.path.join(file_folder_path, file)
        if os.path.exists(file_path): # 파일이 확인되면
        
            origin_img = Image.open(file_path) # 파일을 가져와서
            img = ImageOps.exif_transpose(origin_img)

                # 여기서 순서 안바뀌나 확인할것
            bounding_box = resized_df[resized_df['file_name'] == file]

            print(file)
            x0 = bounding_box['resize_x_center'].values[0] - bounding_box['resize_width'].values[0] / 2
            x1 = bounding_box['resize_x_center'].values[0] + bounding_box['resize_width'].values[0] / 2
            y0 = bounding_box['resize_y_center'].values[0] - bounding_box['resize_height'].values[0] / 2
            y1 = bounding_box['resize_y_center'].values[0] + bounding_box['resize_height'].values[0] / 2
            

            start_point = (int(x0), int(y0))
            end_point = (int(x1), int(y1))
            image = cv2.rectangle(np.array(img), start_point, end_point, color=(0, 255, 0), thickness=1)
            image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
            cv2.imwrite(f"./box/{num}.jpg", image)
            time.sleep(0.5)
