==> Sentiment/4. labeing&relative_coor.ipynb <==
import pandas as pd
import os

cs=pd.read_csv('./resized_df.csv')
cs
dd=cs[['resize_x_center','resize_y_center','resize_width','resize_height']]/640


merged_df = pd.merge(cs['file_name'], dd, left_index=True, right_index=True)
merged_df

emolist=os.listdir('./수정본')
emolist

file_name=[]
file_emo=[]
for emo in emolist:
    file_list_path=os.path.join('./수정본',emo)
    file_list=os.listdir(file_list_path)
    for file in file_list:
        file_name.append(file)
        file_emo.append(emo)

emolabel={'기쁨':0, '당황':1, '분노':2, '불안':3, '상처':4, '슬픔':5, '중립':6}

file_emo = [emolabel[i] for i in file_emo]
file_emo

namedf=pd.DataFrame({'file_name':file_name,'file_emo':file_emo})
namedf

merged_dfa = pd.merge(namedf,merged_df)
merged_dfa

merged_dfa.to_csv('./label_reled.csv',index=False)
