==> Sentiment/5. txt_generator.ipynb <==
import pandas as pd
import os

labels=pd.read_csv('./label_reled.csv')
labels

file_path

ttvlist=os.listdir('./image')
for ttv in ttvlist:
    ttv_folder=os.path.join('./image',ttv)
    emolist=os.listdir(ttv_folder)
    for emo in emolist:
        emofolder=os.path.join(ttv_folder,emo)
        file_list=os.listdir(emofolder)
        for file in file_list:
            file_path=os.path.join(emofolder,file)
            # result = labels[labels['file_name'] == file]
            result = labels[labels['file_name'].str[60:] == file]

            
                    
            file_emo=result['file_emo'].values[0]
            resize_x_center=result['resize_x_center'].values[0]
            resize_y_center=result['resize_y_center'].values[0]
            resize_width=result['resize_width'].values[0]
            resize_height=result['resize_height'].values[0]
             
            
            save_dir=os.path.join('./분할라벨',ttv,emo,file)
            if not os.path.exists(save_dir): #경로가 없으면
                os.makedirs(os.path.dirname(save_dir),exist_ok=True) # 경로를 생성한다. 있으면 무시
            elif os.path.exists(save_dir):
                pass
            
            with open(save_dir+f'{file}.txt', 'w') as file:
                file.write(str(file_emo))
                file.write(str(' '))
                file.write(str(resize_x_center))
                file.write(str(' '))
                file.write(str(resize_y_center))
                file.write(str(' '))
                file.write(str(resize_width))
                file.write(str(' '))
                file.write(str(resize_height))
                file.write(str('\n'))
            
            

            

result = labels[labels['file_name'] == file]

# 결과 출력
print(result['resize_x_center'].values[0])
