==> Sentiment/6. file_numnering.ipynb <==
폴더 스플릿

import splitfolders
splitfolders.ratio('./원본', output='data', seed=77, ratio=(0.6, 0.2, 0.2), group_prefix='data_')


import os

폴더내부 파일 이름순 숫자로 바꾸기

dir='./data'

vvtlist=os.listdir(dir)
for vvt in  vvtlist:
    image_label_folder=os.path.join(dir,vvt)
    label_image=os.listdir(image_label_folder)

    for im in label_image:
        file_folder=os.path.join(image_label_folder,im)
        file_list=os.listdir(file_folder)
        i=0
        for file in file_list:
            i=i+1
            file_path=os.path.join(file_folder,file)

            os.rename(file_path,file_folder+'/'+str(i)+file[-4:])




file_folder+'/'+str(i)+file[-4:]
